Move more files around.
