+++ packages/core/getAggregatedAvailability/date-range-utils/mergeOverlappingDateRanges.test.ts
@@ -3,7 +3,7 @@ import { describe, it, expect } from "vitest";
 import dayjs from "@calcom/dayjs";
 import type { DateRange } from "@calcom/lib/date-ranges";
 
-import { mergeOverlappingDateRanges } from ".";
+import { mergeOverlappingDateRanges } from "./mergeOverlappingDateRanges";
 
 const november2 = "2023-11-02";
 const november3 = "2023-11-03";
