+++ packages/core/getAggregatedAvailability/date-range-utils/mergeOverlappingDateRanges.ts
